--- pkg/reconciler/instances/busola-migrator/types.go
@@ -1,3 +1,0 @@
-package busola_migrator
-
-
