--- pkg/reconciler/instances/busola-migrator/vs_client.go
@@ -1,5 +1,0 @@
-package busola_migrator
-
-type VirtualServiceClient struct {
-
-}
