+++ pkg/reconciler/instances/busola/pre_install_test.go
@@ -1,10 +1,9 @@
-package busola_migrator
+package busola
 
 import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/kyma-incubator/reconciler/pkg/reconciler/kubernetes"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
