+++ pkg/reconciler/instances/busola/test_files/virtual-service.yaml
